refractor: put objectTree item functions in their own file. and make treeCompoenent receive itemFunctions for its use.
